+++ src/renderer/mixins/openExternal.js
@@ -1,0 +1,14 @@
+import { shell } from 'electron';
+
+export default {
+  methods: {
+    /**
+     * Open a url in the default system browser
+     *
+     * @param {string} url The url to open
+     */
+    openExternal(url) {
+      shell.openExternal(url);
+    },
+  },
+};